Wrap drop table episode_queues in clause becuase it doesn't always exist

+++ db/migrate/20140416003858_remove_episode_queues.rb
@@ -1,5 +1,7 @@
 class RemoveEpisodeQueues < ActiveRecord::Migration
   def change
-    drop_table :episode_queues
+    if ActiveRecord::Base.connection.table_exists? :episode_queues
+      drop_table :episode_queues
+    end
   end
 end
